refactor(chat): extract model and response controls
